[FAB-4010] Update nodejs example for 1.0.0-alpha2
Aside from the obvious "it works with alpha2", other
notable changes are that we dropped the fabric-ca-client.

Since 1.0 doesn't support TCerts/Attributes, the CA
was only adding complexity that we couldn't capitalize
on.  Therefore, we now switch to a model where the
ECert is imported directly, rather than translating
a username/password into an ECert via the CA.

Fixes FAB-4010

Change-Id: I753e174d260ad06d101800a1d9cfe7bdfd290aa7
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>