[FAB-9606] Convert indents to TAB in golang.stg
This patch converts all indent whitespace to use the gofmt standard
of TAB.  It also sets an emacs local variable to enable go-mode
when this file is edited by emacs, helping to ensure that the
indent style stays consistent in the future.

There should be no logical changes and no impact to emitted code
(since chaintool runs the output through gofmt anyway).  The main
benefit is in the maintenance of the template.

Fixes FAB-9606

Change-Id: I90e97c9ba1b00547769fd2695f3784974228677d
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>