[FABCT-53] Fix support for referencing embedded types
We add support for referencing types nested within other
definitions.  There were two primary problems:

1) Our EBNF was capturing nested notation as a vector,
e.g. "Foo.MyEnum" was captured as ["Foo" "." "MyEnum"]

2) Our AST validation logic only was designed to handle
inner-to-outer scope.  For the cases where the CCI was
defined to utilize nested messages, our validator needs
to understand to look in both directions (e.g. including
outer-to-inner).  We now define the former as "local scope"
and the latter as "global" scope, and include both dimensions
in the validator.

Fixes FABCT-53

Change-Id: Iba87dfd27b9d65481fb04e0ce993c1c5180deaa9
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>